Add playGame() and rework some logic
playGame initializes scores for the player and the computer.
then begins a loop that reinitializes player and computer
choices each time it loops. If player choice is the same
as computer, loop skips over the rest and begins again
at the loop conditional. if choices are different, they will
be compared to see if its a win for the player or not. the
return of playRound() will determine the outcomes of the scores.
First to 5 wins the game.